LayoutTests/crypto/webkitSubtle/aes-cbc-import-jwk-expected.txt

CONSOLE MESSAGE: line 32: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test importing a JWK key for AES-CBC.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...

PASS key.type is 'secret'
PASS key.extractable is false
PASS key.algorithm.name is 'AES-CBC'
PASS key.algorithm.length is 192
PASS key.usages is ["encrypt"]

Using the key to encrypt plaintext...
PASS bytesToHexString(new Uint8Array(encryptionResult)) is '4f021db243bc633d7178183a9fa071e8b4d9ada9ad7dedf4e5e738763f69145a571b242012fb7ae07fa9baac3df102e008b0e27988598881d920a9e64f5615cd612ccd79224b350935d45dd6a98f8176'
PASS successfullyParsed is true

TEST COMPLETE
